feat: add Guide

+++ database/migrations/2024_06_15_202452_create_budgets_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('budgets', function (Blueprint $table) {
             $table->id();
             $table->foreignIdFor(\App\Models\User::class);
-            $table->double('budget_amount', 13, 2);
+            $table->float('budget_amount', 13, 2);
             $table->timestamps();
         });
     }
